Updates for the Jest framework.
