+++ src/polyfills.ts
@@ -1,4 +1,2 @@
-// Used for reflect-metadata in JIT. If you use AOT (and only Angular decorators), you can remove.
-import 'core-js/es7/reflect';
-// This is used to polyfill on IOS - used for detecting when an image comes into view
 import 'zone.js/dist/zone'; // Included with Angular CLI.
+import './zone-flags.ts';
