+++ src/tsconfig.spec.json
@@ -11,8 +11,11 @@
       "node"
     ]
   },
+  "files": [
+    "zone-flags.ts",
+    "polyfills.ts"
+  ],
   "include": [
-    "polyfills.ts",
     "**/*.spec.ts",
     "**/*.d.ts"
   ]
