+++ src/zone-flags.ts
@@ -1,0 +1,5 @@
+/**
+ * Prevents Angular change detection from
+ * running with certain Web Component callbacks
+ */
+(window as any).__Zone_disable_customElements = true;
