+++ tsconfig.json
@@ -15,7 +15,12 @@
     "paths": {
       "app/*": ["./app/*"],
       "environments/*": ["./environments/*"],
-      "lib/*": ["./lib/*"]
+      "lib/*": ["./lib/*"],
+      "core-js/es7/reflect": [
+        "node_modules/core-js/proposals/reflect-metadata"
+      ],
+      "core-js/es6/*": ["node_modules/core-js/es"],
+      "core-js/es7/*": ["node_modules/core-js/es"]
     },
     "removeComments": true,
     "sourceMap": true,
